refactor: move behavior from `mls::proposal` module to `transaction_context` module
